[#19]Refactor: 인터페이스 지정 및 함수 추출
